backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
langchain==0.2.16
langchain-community==0.2.17
langchain-openai==0.1.23
sqlalchemy==2.0.35
google-generativeai==0.7.2